refactor(crm): run BullMQ consumers in-process, drop separate worker entry
The repo never had a worker process before slice 1 — moving the queue
consumers back into the API process so the runtime layout matches what
existed on main + what existed before the lift.

- src/index.ts now starts crmEmailBackfillWorker + crmStageClassifyWorker
  + the reclassify-on-deploy one-shot hook AFTER the API server begins
  listening. Adds SIGTERM/SIGINT shutdown that closes workers, queues,
  and the Redis connection before exit.
- Deletes src/worker.ts (the separate-process entry point from Phase 0).
- Reverts package.json dev/start scripts to single-process bun.
- Removes the concurrently dev dep.
- Tweaks .env.example REDIS_URL comment to reflect "consumers in-process".
