+++ elysia-server/.env.example
@@ -79,10 +79,11 @@ MIXPANEL_PROJECT_ID=              # [OPTIONAL]
 MIXPANEL_PROJECT_SECRET=          # [OPTIONAL]
 
 # -----------------------------------------------------------------------------
-# BullMQ / Redis  [REQUIRED for the worker process]
+# BullMQ / Redis  [REQUIRED]
 # -----------------------------------------------------------------------------
-# Used by CRM email-backfill and stage-classify workers. `docker compose -f
-# compose.db.yml up -d` brings up a local Redis on 6379.
+# Used by CRM email-backfill and stage-classify queues; consumers run
+# in-process alongside the API. `docker compose -f compose.db.yml up -d`
+# brings up a local Redis on 6379.
 REDIS_URL=redis://localhost:6379
 
 # -----------------------------------------------------------------------------
