+++ elysia-server/package.json
@@ -4,11 +4,8 @@
   "description": "RINDA CRM Backend - TypeScript/Elysia/PostgreSQL",
   "type": "module",
   "scripts": {
-    "dev": "concurrently -n api,worker -c blue,green \"bun run --watch src/index.ts\" \"bun run --watch src/worker.ts\"",
-    "dev:api": "bun run --watch src/index.ts",
-    "dev:worker": "bun run --watch src/worker.ts",
+    "dev": "bun run --watch src/index.ts",
     "start": "bun run src/index.ts",
-    "start:worker": "bun run src/worker.ts",
     "db:generate": "drizzle-kit generate",
     "db:push": "drizzle-kit push",
     "db:studio": "drizzle-kit studio",
@@ -57,7 +54,6 @@
     "@types/marked": "^6.0.0",
     "@types/pg": "^8.15.5",
     "bun-types": "^1.3.1",
-    "concurrently": "^9.2.1",
     "drizzle-kit": "^0.30.0",
     "typescript": "^5.7.0"
   }
